Refactoring: Slim fat module
